+++ crates/slipbox-rpc/src/lib.rs
@@ -8,6 +8,7 @@ pub const METHOD_PING: &str = "slipbox/ping";
 pub const METHOD_INDEX: &str = "slipbox/index";
 pub const METHOD_INDEX_FILE: &str = "slipbox/indexFile";
 pub const METHOD_SEARCH_NODES: &str = "slipbox/searchNodes";
+pub const METHOD_RANDOM_NODE: &str = "slipbox/randomNode";
 pub const METHOD_SEARCH_TAGS: &str = "slipbox/searchTags";
 pub const METHOD_NODE_FROM_ID: &str = "slipbox/nodeFromId";
 pub const METHOD_NODE_FROM_TITLE_OR_ALIAS: &str = "slipbox/nodeFromTitleOrAlias";